update:Scene0にマーカごとの距離取得機能を追加、その他コメントの修正など
Scene0にスタートからゴールまでの各マーカー間の距離をmarker_distance[]に記録する機能を追加（配列は100個までなので、コースによってはオーバーする可能性あり）
Logにマーカのカウントなどを出力するように修正

+++ soucecode/picotracer/types.h
@@ -6,9 +6,9 @@
     //ライン検知センサ
     uint16_t Curve_log;
     uint16_t LL_log;
-    uint16_t L_log;
+    uint32_t L_log;
     uint16_t R_log;
-    uint16_t RR_log;
+    uint32_t RR_log;
     int Goal_log;
     //モーター出力
 //    uint16_t R_motor_log;
